Add aggregated ESP32 reading uploads

+++ ESP32C3/src/pms5003.h
@@ -2,12 +2,12 @@
 #include <Arduino.h>
 
 struct PMSdata{
-    int pm1_0; //concetration of pms1.0 particles
-    int pm2_5; //concentration of pms 2.5 particles 
-    int pm10_0; //concentraton of pms 10 particles
+    uint16_t pm1_0; //concetration of pms1.0 particles
+    uint16_t pm2_5; //concentration of pms 2.5 particles
+    uint16_t pm10_0; //concentraton of pms 10 particles
 };
 
-void pmsInit(); 
+void pmsInit();
 
 bool pmsRead(PMSdata &data);
